test(inbound): split the preservation retry cases out of a 1089-line file
A maintainability review flagged tests/test_inbound_preservation.py crossing
the 1000-line limit once the retry cases landed.

The retry tests and their retry-specific helpers move to
tests/test_inbound_preservation_retry.py. The setup both files need -- the
connection seed and the queued-submission assertion -- moves to
tests/inbound_preservation_support.py, following the repo's existing
tests/scheduler_test_support.py and tests/slack_monitor_fixtures.py
convention.

No test is added, removed, renamed or re-parametrized: the collect-only count
for preservation_setup_retry is 32 before and after, with matching ids. The
retry tests now import run_tool_execution_started directly from
brain.systems.runs.failure_diagnostic rather than through the inbound service
module.

1089 -> 723 / 265 / 138 lines. No file under brain/ changes.
